updated for refactoring
